Simplify build process
- Originally I started out using a Portico style ant build for SysCommon.
  While this is very powerful it does not support static linking of the
  output targets, which is the main use case for SysCommon at the moment.
- As a result I'm temporarily moving to batch/shell script based compiling
  because its simple and easy.
- This has the additional benefits of including Visual Studio projects for
  the codebase in the main distro which makes the development setup
  process a bit nicer
- Also fixed unit tests for the previous namespace/structure changes

+++ codebase/src/cpp/syscommon/src/OutputBuffer.cpp
@@ -84,7 +84,7 @@ void OutputBuffer::writeInt64( long long value ) throw ( IOException )
 void OutputBuffer::writeUTF( const std::string& value ) throw ( IOException )
 {
 	size_t length = value.length();
-	this->writeUInt16( length );
+	this->writeUInt16( (unsigned short)length );
 
 	if( this->getBytesRemaining() >= length )
 	{
